Update affected user flow to not use modal mode
Why: prior to this commit, the affected user flow stack was being
displayed within a modal using React Navigation. However, it is not
possible (as of now) to disable the "swipe down to close" gesture for
modals presented via React Navigation. Given that we want to disable
this gesture, we had to take a different approach.

This commit:
- Removes the modal mode from the options for the affected user flow
  stack
- Instead uses a custom cardStyleInterpolator to achieve the
  animateFromBottom transition when opening this stack